Deployment: Allow non-default docker hosts
The setup uses the IP address of the host to resolve some names.
When docker is configured to use a different range of IP addresses,
this stops working correctly.

The main use case for this is that the WiFi on DB ICE trains allocates
IPs from the 172.17.0.0/16 range, which is the same as the default
docker network. When both use the same range, problems occur with both
docker and the WiFi.