Keycloak: Add example for using friendlycaptcha for registration and login.
- Enable friendlycaptcha
To use friendlycaptcha first register an account and create an app at https://friendlycaptcha.com/
After that set friendlyCaptchaEnabled: "true".
- Captcha for Registration
In acme-apps the registration can be protected with an additional friendly captcha. To enable the captcha just enable the "acme-friendly-captcha-form-action" in the custom Registration Flow.
- Captcha for Login
Replace UsernamePasswordForm authenticator with "Acme: Captcha Username Password Form" in a custom browser flow.